% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_neighbours}
\alias{add_neighbours}
\title{Add neighbouring parameter values}
\usage{
add_neighbours(scenarios, parameter, radius, points, drop = T)
}
\arguments{
\item{scenarios}{matrix with one parameter setting in each line}

\item{parameter}{list of parameters for which grid points should be added}

\item{radius}{width of interval to be added around each point in \code{scenarios}}

\item{points}{number of points to fill the interval with}

\item{drop}{should columns that start with \code{"result"} or \code{"select"} or  be dropped (which may have been added by \code{\link{simulate_batch}}, or \code{\link{select_results}})}
}
\value{
object with the same columns as G
}
\description{
Expand and refine parameter grid for simulation
}
\examples{
scenarios <- expand.grid(m=c(0,1),s=c(1,2),n=5:10)
neighbourhoods <- add_neighbours(scenarios,c('m','s'),c(.2,.4),4)

}
\author{
Florian Klinglmueller
}

