% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blackjack.R
\docType{data}
\name{Blackjack}
\alias{Blackjack}
\title{Blackjack R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Blackjack
}
\description{
Blackjack R6 Class
}
\keyword{datasets}
