% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deck.R
\docType{data}
\name{Deck}
\alias{Deck}
\title{Deck R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Deck
}
\description{
Deck R6 Class
}
\examples{
# create a new deck
x <- Deck$new()
x

# draw a card
x$draw(1)
x

# draw 10 cards
x$draw(10)

# check how many cards are left
x$cards_left()

# reset the deck
x$shuffle()
x

# create a deck composed of 5 decks
x <- Deck$new(decks = 5)
x
}
\keyword{datasets}
