% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_sound.R
\name{play_sound}
\alias{play_sound}
\title{Play a sound (if possible)}
\usage{
play_sound(sound = "fanfare")
}
\arguments{
\item{sound}{character string or number specifying the sound (see \code{\link[beepr]{beep}})}
}
\description{
Play a sound (if possible)
}
\note{
requires the `beepr` package
}
