% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{players}
\alias{players}
\title{List all player profiles}
\usage{
players(file = Sys.getenv("CASINO_FILE"))
}
\arguments{
\item{file}{full path to file containing player profiles}
}
\description{
List all player profiles
}
