% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_scenarios}
\alias{plot_scenarios}
\title{plot scenarios}
\usage{
plot_scenarios(old, selected, new, parameter, groups)
}
\arguments{
\item{old}{Original scenarios from which selection was made}

\item{selected}{Selected scenarios}

\item{new}{New scenarios based on selection}

\item{parameter}{Nuisance parameter}

\item{groups}{Parameter of interest}
}
\value{
scatter plot of scenarios
}
\description{
Plot old, selected and new candidate scenarios.
}
\examples{
scenarios <- expand.grid(m=seq(0,1,.01),s=seq(1,2,.01),n=1000)
results <- simulate_batch(scenarios,100000,simulate_poly)
candidates <- select_results(results,'s',"result")
smooth <- smooth_loess(candidates,'m','s',10)
grid <- add_neighbours(candidates,'m',.2,10)
plot_scenarios(results,candidates,smooth,'m','s')
plot_scenarios(results,candidates,grid,'m','s')
}
\author{
Florian Klinglmueller
}

