% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poker.R
\docType{data}
\name{Poker}
\alias{Poker}
\title{Poker R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Poker
}
\description{
Poker R6 Class
}
\examples{
set.seed(101315)
setup()

# draw poker
x <- Poker$new(who = "Player 1", type = "draw", bet = 10)
x$play()
x$hold(1, 2, 5)
x$draw()

# stud poker (bet 20)
x <- Poker$new(who = "Player 1", type = "stud", bet = 20)
x$play()

# clean-up
delete()
}
\keyword{datasets}
