% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_results}
\alias{select_results}
\title{Select results}
\usage{
select_results(sim_results, group, what, functional = which.max)
}
\arguments{
\item{sim_results}{Simulation results from a previous simulation run}

\item{group}{Grouping variables within unique values of which results should be selected}

\item{what}{Variable based on which the selection is performed}

\item{functional}{Function that returns a unique vector position}
}
\value{

}
\description{
This function selects rows from a dataframe of simulation results. For example using the default \code{functional} \code{which.max} the rows containing the maximum of column \code{what} within values of \code{group} are selected. The function adds column with name \code{select_what} that specifies based on which columns a row was selected.
}
\examples{
scenarios <- expand.grid(m=c(0,1),s=c(1,2),n=10:20)
results <- simulate_batch(scenarios,10000,simulate_poly)
select_results(results,'n',"result")
}
\author{
Florian Klinglmueller
}

