% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Allow casino to store player profiles in a local file (and)}
\usage{
setup(file = file.path(getwd(), ".casino"))
}
\arguments{
\item{file}{full path to file}
}
\description{
Allow casino to store player profiles in a local file (and)
}
