% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_batch}
\alias{simulate_batch}
\title{Run a batch of simulations}
\usage{
simulate_batch(scenarios, runs, run_scenario, ..., use_mclapply2 = FALSE,
  multicore = TRUE)
}
\arguments{
\item{scenarios}{Object with paramter settings for \code{delta}, \code{sigma} and \code{n1} in each row}

\item{runs}{Number of simulation runs for each scenario}

\item{run_scenario}{Function that runs a single parameter scenario, takes at least the columns of \code{G} and \code{runs} as arguments}

\item{...}{Additional named arguments for run_scenario}

\item{use_mclapply2}{use mclapply2 from package bt88.03.704 which implements a status bar (see details)}

\item{multicore}{Should multicore parallelisation be used}
}
\value{
Object with parameters and simulation results in each row
}
\description{
To use mclapply2 from bt88.03.704 you have to install this library (e.g. \code{devtools::install_github('floatofmath/bt88.03.704')}) unfortunately the statusbar implementation does not work with Rstudio.
}
\examples{

scenarios <- expand.grid(m=c(0,1),s=c(1,2))
simulate_batch(scenarios,10000,simulate_poly,n=10)
}
\author{
Florian Klinglmueller
}

