% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_poly}
\alias{simulate_poly}
\title{Disturbed function}
\usage{
simulate_poly(m, s, n, runs)
}
\arguments{
\item{m}{parameter \code{m}}

\item{s}{parameter \code{s}}

\item{n}{parameter \code{s}}

\item{runs}{parameter \code{runs}}
}
\value{
numeric value of the disturbed function
}
\description{
A strange function with error to maximize
}
\examples{
par(mfrow=c(4,1))
plot(x <- seq(0,1,.001),simulate_poly(x,1,100,10^6))
plot(x <- seq(0,1,.001),simulate_poly(x,1.3,100,10^6))
plot(x <- seq(0,1,.001),simulate_poly(x,1.6,100,10^6))
plot(x <- seq(0,1,.001),simulate_poly(x,2,100,10^6))
}
\author{
Florian Klinglmueller
}

