% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smooth_loess}
\alias{smooth_loess}
\title{smooth parameter grid}
\usage{
smooth_loess(scenarios, parameter, groups, points, drop = T, width = 6, ...)
}
\arguments{
\item{scenarios}{list of candadidate parameter settings}

\item{parameter}{parameter to smooth}

\item{groups}{explanatory variable}

\item{points}{number of points around each prediction}

\item{drop}{should columns that start with \code{"result"} or \code{"select"} or  be dropped (which may have been added by \code{\link{simulate_batch}}, or \code{\link{select_results}})}

\item{width}{width of the band around the smoothed fit in multiples of the standard error}

\item{...}{additional options to \code{\link{stats::loess}}}
}
\value{
matrix of parameter settings with \code{points} times \code{nrow(scenarios)} rows
}
\description{
Fits a loess regression and adds candidates scenarios within the 95 percent confidence band around the fit.
}
\examples{
scenarios <- expand.grid(m=seq(0,1,.1),s=seq(1,2,.1),n=100)
results <- simulate_batch(scenarios,100000,simulate_poly)
candidates <- select_results(results,'s',"result")
smooth_criminals <- smooth_loess(candidates,'m','s',10)

}
\author{
Florian Klinglmueller
}

