% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blackjack.R
\docType{data}
\name{Blackjack}
\alias{Blackjack}
\title{Blackjack R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Blackjack
}
\description{
Blackjack R6 Class
}
\examples{
set.seed(101315)
setup()

# sit at the blackjack table
x <- Blackjack$new(who = "Player 1", bet = 10)

# play a hand
x$play()

x$hit()

x$stand()

# play a hand blind w/out drawing
x$play()$stand()

# clean-up
delete()
}
\keyword{datasets}
